Numeric.Hamilton

Documentation

data System :: Nat -> Nat -> Type

Represents a physical system in which physics happens. A System m n is a system whose state described using n generalized coordinates (an "n-dimensional" system), where the underlying cartesian coordinate space is m-dimensional.

For the most part, you are supposed to be able to ignore m. m is only provided because it's useful when plotting/drawing the system with a given state back in rectangular coordinates.

A System m n's state is described using a Config n (which describes the system in configuration space) or a Phase n (which describes the system in phase space).

data Config :: Nat -> Type where

Represents the full state of a system of n generalized coordinates in configuration space (informally, "positions and velocities")

A configuration space representaiton is more directly "physically meaningful" and intuitive/understandable to humans than a phase space representation. However, it's much less mathematically ideal to work with because of the lack of some neat underlying symmetries.

You can convert a Config n into a Phase n (convert from configuration space to phase space) for a given system with toPhase. This allows you to state your system in configuration space and then convert it to phase space before handing it off to the hamiltonian machinery.

Constructors

Cfg :: {..} -> Config n

Fields

cfgPos :: !(R n)

The current values ("positions") of each of the n generalized coordinates

cfgVel :: !(R n)

The current rate of changes ("velocities") of each of the n generalized coordinates

data Phase :: Nat -> Type where

Represents the full state of a system of n generalized coordinates in phase space (informally, "positions and momentums").

Phase space representations are much nicer to work with mathematically because of some neat underlying symmetries. For one, positions and momentums are "interchangeable" in a system; if you swap every coordinate's positions with their momentums, and also swap them in the equations of motions, you get the same system back. This isn't the case with configuration space representations.

A hamiltonian simulation basically describes the trajectory of each coordinate through phase space, so this is the state of the simulation. However, configuration space representations are much more understandable to humans, so it might be useful to give an initial state in configuration space using Config, and then convert it to a Phase with toPhase.

Constructors

Phs :: {..} -> Phase n

Fields

phsPos :: !(R n)

The current values ("positions") of each of the n generalized coordinates.

phsMom :: !(R n)

The current conjugate momenta ("momentums") to each of the n generalized coordinates

mkSystem

Arguments

:: (KnownNat m, KnownNat n)

=> R m
The "inertia" of each of the m coordinates in the underlying cartesian space of the system. This should be mass for linear coordinates and rotational inertia for angular coordinates.

-> (forall a. RealFloat a => Vector n a -> Vector m a)
Conversion function to convert points in the generalized coordinate space to the underlying cartesian space of the system.

-> (forall a. RealFloat a => Vector m a -> a)
The potential energy of the system as a function of points in the underlying cartesian space of the system.

-> System m n

Create a system with n generalized coordinates by providing the underlying inertials, describing its coordinate space, and giving pontential energy function.

Given by describing some fundamental properties of the underlying cartesian coordinate space of the system (the inertia of each coordinate and the potential energy function), along with a conversion function from the generalized coordinate space to the underlying cartesian coordinate space.

underlyingPos :: System m n -> R n -> R m

Converts the position of generalized coordinates of a system to the coordinates of the system's underlying cartesian coordinate system. Useful for plotting/drawing the system in cartesian space.

pe :: System m n -> R n -> Double

The potential energy of a system, given the position in the generalized coordinates of the system.

toPhase :: (KnownNat m, KnownNat n) => System m n -> Config n -> Phase n

Convert a configuration-space representaiton of the state of the system to a phase-space representation.

Useful because the hamiltonian simulations use Phase as its working state, but Config is a much more human-understandable and intuitive representation. This allows you to state your starting state in configuration space and convert to phase space for your simulation to use.

fromPhase :: (KnownNat m, KnownNat n) => System m n -> Phase n -> Config n

Invert toPhase and convert a description of a system's state in phase space to a description of the system's state in configuration space.

Possibly useful for showing the phase space representation of a system's state in a more human-readable/human-understandable way.

momenta :: (KnownNat m, KnownNat n) => System m n -> Config n -> R n

Compute the generalized momenta conjugate to each generalized coordinate of a system by giving the configuration-space state of the system.

velocities :: (KnownNat m, KnownNat n) => System m n -> Phase n -> R n

Compute the rate of change of each generalized coordinate by giving the state of the system in phase space.

keC :: (KnownNat m, KnownNat n) => System m n -> Config n -> Double

The kinetic energy of a system, given the system's state in configuration space.

keP :: (KnownNat m, KnownNat n) => System m n -> Phase n -> Double

The kinetic energy of a system, given the system's state in phase space.

lagrangian :: (KnownNat m, KnownNat n) => System m n -> Config n -> Double

The Lagrangian of a system (the difference between the kinetic energy and the potential energy), given the system's state in configuration space.

hamiltonian :: (KnownNat m, KnownNat n) => System m n -> Phase n -> Double

The Hamiltonian of a system (the sum of kinetic energy and the potential energy), given the system's state in phase space.

hamEqs :: (KnownNat m, KnownNat n) => System m n -> Phase n -> (R n, R n)

stepHam :: (KnownNat m, KnownNat n) => Double -> System m n -> Phase n -> Phase n

evolveHam

Arguments

:: (KnownNat m, KnownNat n, KnownNat s)

=> System m n
system to simulate

-> Phase n
initial state, in phase space

-> Vector s Double
desired solution times

-> Vector s (Phase n)

Evolve a system using a hamiltonian stepper, with the given initial phase space state.

evolveHam' :: forall m n. (KnownNat m, KnownNat n) => System m n -> Phase n -> [Double] -> [Phase n]

Evolve a system using a hamiltonian stepper, with the given initial phase space state.

Desired solution times provided as a list instead of a sized Vector. The output list should be the same length as the input list.

evolveHamC

Arguments

:: (KnownNat m, KnownNat n, KnownNat s)

=> System m n
system to simulate

-> Config n
initial state, in configuration space

-> Vector s Double
desired solution times

-> Vector s (Config n)

A convenience wrapper for evolveHam that works on configuration space states instead of phase space states.

Note that the simulation itself still runs in phase space; this function just abstracts over converting to and from phase space for the inputs and outputs.

evolveHamC'

Arguments

:: (KnownNat m, KnownNat n)

=> System m n
system to simulate

-> Config n
initial state, in configuration space

-> [Double]
desired solution times

-> [Config n]

A convenience wrapper for evolveHam' that works on configuration space states instead of phase space states.

Note that the simulation itself still runs in phase space; this function just abstracts over converting to and from phase space for the inputs and outputs.